move tests to more appropriate file
